+++ lib/broadside/configuration/verify_instance_variables.rb
